backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.21
pydantic-settings==2.3.1
python-multipart==0.0.9
yt-dlp==2024.8.6
faster-whisper==1.0.3
python-pptx==0.6.23
pdfplumber==0.11.0
chromadb==0.5.5
transformers==4.51.3
torch==2.3.1+cu121
sentencepiece==0.2.0
accelerate==0.32.1
tokenizers==0.21.4
